Make the protocol uniquing more tolerant when merging protocols.  Fixes #23.
It merges the methods individually, instead of assuming the method count will be either zero or equal.
This allows it to merge an NSObject protocol where -debugDescription is listed as an instance method with
one where it is listed as an optional instance method.

An existing instance method is moved to the optional instance methods, if necessary.

/Applications/Xcode.app/Contents/OtherFrameworks/DevToolsInterface.framework/Versions/A/DevToolsInterface Current version: 2301.0.0
  - duplicate anchorIsEditing: method removed from instance methods
/Applications/Messages.app/Contents/MacOS/Messages Source version: 3312.0.0.0.0
  - duplicate tableView:mouseMoved: removed